@@ -4,7 +4,7 @@ version: 0.0.7
 homepage: https://decodedhealth.com
 issue_tracker: https://github.com/decodedhealth/flutter_zoom_plugin/issues
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.0 <2.0.0"
 
 dependencies:
