@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'dart:async';
 
 import 'package:flutter_zoom_plugin_example/join_screen.dart';
 import 'package:flutter_zoom_plugin_example/meeting_screen.dart';
@@ -10,13 +9,12 @@ void main() => runApp(ExampleApp());
 class ExampleApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-
     return MaterialApp(
       title: 'Example Zoom SDK',
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      navigatorObservers: [ ],
+      navigatorObservers: [],
       initialRoute: '/',
       routes: {
         '/': (context) => JoinWidget(),
